fix(canvas)!: Use full block for Marker::Block
When drawing a vertical line on a canvas, using `Marker::Block` would
draw a gap in between each line. This is because the default block
character was a half bar. This changes the default block character to a
full block, and adds a new marker type `Marker::Bar` to replace the
current `Marker::Block`.

BREAKING CHANGE: `Marker::Block` now uses a full block character, and
`Marker::Bar` is the new half block character.

- Adds tests for each `Marker` in the canvas widget.
- Adds marker to the canvas example to show the difference between
  the marker types.