feat: simplify backend and terminal creation
This commit simplifies the creation of backends and terminals by
implementing `Default` for the backends and adding builder style
methods to the terminal.

For example, the following code now creates a terminal that enters the
alternate screen and enables raw mode and cleans up after itself when
dropped.

```rust
let backend = CrosstermBackend::on_stdout();
let mut terminal = Terminal::new(backend)?;

loop {
    terminal.draw(|f| {
        // draw stuff
    })?;
    ...
}
```

CrosstermBackend and TermionBackend both gain `on_stdout()` and
`on_stderr()` methods which enable raw mode and switch to the alternate
screen before returning, and builder style methods to enable or disable
raw mode and the alternate screen.

This commit implements `Drop` for the CrosstermBackend which will
disable raw mode and switch back to the main screen. This ensures the
terminal is left in a usable state if the application panics. The
underlying Termion and Termwiz code both already implement `Drop` so no
changes are required to them.

This commit adds a Deref implementation to terminal to allow the backend
to be accessed directly.

The README and library, terminal, backends, types, traits and modules
docs are all updated to reflect these changes and to add some missing
documentation and examples.

Fixes https://github.com/tui-rs-revival/ratatui/issues/261