feat(layout): rename SegmentSize to Spread
For some time we've been using the term "segment size" to refer to the
way extra space is distributed when the constraints are satisfied. This
is a very confusing name, and it's not clear what it means. This commit
renames it to "spread", which is a more common term for this concept.

Removed the unstable feature gate for this API now that a good name has
been found.

BREAKING CHANGE:
- SegmentSize has been renamed to Spread
- unstable-segment-size feature has been removed
- Layout::segment_size() has been renamed to Layout::spread()
- Table::segment_size() has been renamed to Table::spread()