feat(table): Implement FromIterator for widgets::Row
The `Row::new` constructor accepts a single argument that implements
`IntoIterator`.  This commit adds an implementation of `FromIterator`,
as a thin wrapper around `Row::new`.  This allows `.collect::<Row>()`
to be used at the end of an iterator chain, rather than wrapping the
entire iterator chain in `Row::new`.

Examples in `examples/table.rs` that used an intermediate variable for
the cell iterator have been updated to instead collect directly into a
`Row`.  Examples that have an explicit `vec![cell1, ..., cellN]` have
been kept as-is.