fix(list): fixed an issue with large padding values and small render areas, added tests for those cases, fixed missing documentation
Previously large padding values could push the selected item out of the viewable area, or could cause it to appear to jump up and down as the new effective offset value was calculated differently on each render. Updated the code to reduce the padding value to make sure it will always fit inside the avalible render area. Added tests to cover both of those potential issues. Updated documentation for ListState that I missed in the first commit for this feature.

Signed-off-by: Cameron Barnes <cameron_barnes@outlook.com>