fix(termion): Add workaround for failing cursor detection
The termion backend now stores a last known cursor position. If termion
fails to get the cursor position, it will return the last known cursor
position, if available. If there is no known cursor position, it will
return an error.

This is a workaround for a when the thread that is responsible for
reading from stdin is not the same as the thread that is responsible for
writing to stdout.