perf(borders): allow border!() in const
This allows more compiler optimizations when the macro is used.

Interestingly TOP | BOTTOM is bitwise or and not const (it's a trait…).
insert is also not const, union is.

Also moves tests from test folder to the file itself as I overlooked
them initially. Borders was together in a file with Block back when it
was introduced.