perf(scrollbar): pass symbol set by reference
A symbol set contains &'static str. It itself only needs to be some reference as whatever is in it has to be static anyway and only the inner values are used after the method ends and the set is out of scope again.

BREAKING CHANGE: `Scrollbar::symbols` used to move the value and now only needs a reference.