docs(readme): remove MSRV
This notice was useful when the Cargo.toml had no standardized field for this. Now its easier to look it up in the Cargo.toml and its also a single point of truth. Updating the README was overlooked for quite some time so its better to just omit it rather than having something wrong that will be forgotten again in the future.