Reimplement Terminal::insert_before.
The main reason to reimplement this function is to fix #999. As an added
bonus, the new version is more efficient when the viewport is large.

The old version would take lines from the buffer to insert in chunks.
These chunks contained as many lines as there were "free" lines on the
screen above the viewport. So, if the screen was 100 lines tall, and the
viewport was 80 lines tall, it would use chunks of 20 lines.

The main problem with this approach is that when the viewport takes up
the whole screen, the chunk size drops to zero, which causes the panic
in #999. But the approach has the problem of doing a lot of extra work
when the viewport is almost as tall as the screen. To insert 20 lines,
on a screen 100 lines tall, with a viewport of 99 lines, would require
20 separate chunks. Processing each chunk involved inserting a bunch of
extra lines at the bottom of the screen before drawing the line to
insert.

The new approach basically chunks as well, but with a chunk size equal
to the screen size.