feat(table)!: add column/cell selection to TableState
BREAKING CHANGE: The serialized output of `TableState` will now also
contain the "selected_column". Any application or library that manually
parses the serialized output of the state will have to update to support
the new field. This change does not affect the deserialization, so
existing states do not have to be manually updated.