Add a new feature: scrolling-regions.
This feature will be used to enable optimized Terminal::insert_before code that
uses terminal scrolling regions.

Hopefully, this feature can be removed and the optimized code can replace the
existing Terminal::insert_before code. However, it remains to be seen if
scrolling regions are supported on all terminals that Ratatui supports.