Shorten TestBackend tests.
This commit does two things.

First, it introduces a new TestBackend::with_lines method and uses that to make
a lot of tests one line shorter. More importantly, it removes the hard-coding
of the width and heigh, which then have to match the buffer provided in the
next line.

Second, use rstest to collapse all of the scroll_region_up and
scroll_region_down tests into two tests. This relies on the new
TestBackend::with_lines method.